requirements.txt
africastalking==2.0.1
alembic==1.17.0
annotated-types==0.7.0
any==0.0.2
anyio==4.11.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
fastapi==0.117.1
filelock==3.19.1
Flask==3.1.2
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
idna==3.10
influxdb==5.3.2
influxdb-client==1.49.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
msgpack==1.1.1
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.1.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
pyngrok==7.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.49.2
reactivex==4.0.4
referencing==0.36.2
requests==2.32.5
responses==0.25.8
rpds-py==0.27.1
schema==0.7.7
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.48.0
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
wcwidth==0.2.14
Werkzeug==3.1.3
